agent_bar_v2/subagents/retail/customer_support_hub/pyproject.toml
[project]
name = "customer-support-hub"
version = "0.0.1"
description = "Combined agent pack for Customer Support and Product Search."
authors = [{ name = "example", email = "example@google.com" }]
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.13",
]
keywords = ['agent-packs', 'adk', 'agent']
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.12.0",
    "uvicorn>=0.34.3",
    "python-dotenv",
    "authlib==1.6.4"
]

[build-system]
requires = [
  "setuptools >= 40.9.0",
]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipython>=8.36.0",
    "pytest>=8.3.5",
    "pre-commit>=4.2.0",
    "ruff>=0.11.5",
    "mypy>=1.16.0",
    "deptry>=0.23.0",
    "pytest-cov>=4.0.0",
    "pylint>=3.3.7",
    "google-adk[eval]>=1.3.0",
    "basedpyright"
]

[tool.agent_packs]
template = "adk"

[tool.mypy]
files = ["src"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true
ignore_missing_imports = true

[tool.basedpyright]
reportAny = false
reportExplicitAny = false
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownVariableType = false

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
fix = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["src"]
omit = ["server.py"]

[tool.deptry.per_rule_ignores]
DEP002 = ["uvicorn", "authlib"]
